% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xbiome_jw_function.R
\name{advanced_permanova_test}
\alias{advanced_permanova_test}
\title{Permanova Test with TWO variabl}
\usage{
advanced_permanova_test(d, meta, feature1, feature2, metric)
}
\arguments{
\item{d}{OTU TABLE with rows being sample and columns being features}

\item{meta}{metadata}

\item{feature1}{a column name in meta}

\item{feature2}{a column name in meta}
}
\value{
a statistac res, p value is get by res$aov.tab$`Pr(>F)`[1]
}
\description{
Permanova test used to test the dispersion difference FOR two variables
}
\author{
Wei Jiang <jiangwei@xbiome.cn>
}
