% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xbiome_jw_function.R
\name{bray}
\alias{bray}
\title{Bray-Curtis Distance of Samples}
\usage{
bray(x)
}
\arguments{
\item{x}{OTU table, with rows being samples and columns being features}
}
\value{
distance matrix
}
\description{
Give the OTU table, return the bray-curtis distance matrix
}
\author{
Wei Jiang <jiangwei@xbiome.cn>
}
