% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xbiome_jw_function.R
\name{cosine}
\alias{cosine}
\title{Cosine Distance of Samples}
\usage{
cosine(x)
}
\arguments{
\item{x}{OTU table, with rows being samples and columns being features.}
}
\value{
distance matrix
}
\description{
Give the OTU table, return the cosine distance matrix;
}
\author{
Wei Jiang <jiangwei@xbiome.cn>
}
