% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xbiome_jw_function.R
\name{glmnet_model}
\alias{glmnet_model}
\title{Glmnet Model}
\usage{
glmnet_model(train, test, n = 100, method = "cv", fold = 3)
}
\arguments{
\item{train}{train SET}

\item{test}{test SET}

\item{n}{seed}

\item{method}{method for tuning hyperprameters}

\item{fold}{the number of folds used for cross validation}

\item{method}{method for tuning hyperprameters}
}
\value{
A list containing model, confusion matrix, test AUC and variables used
}
\description{
Apply Generalized Linear Model for trainning in train set and test in test set
}
\author{
Wei Jiang <jiangwei@xbiome.cn>
}
