% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xbiome_jw_function.R
\name{imp_select}
\alias{imp_select}
\title{Feature Selection by Importance}
\usage{
imp_select(dataset, n, top_n = 10)
}
\arguments{
\item{dataset}{Train Dataset}

\item{n}{seed}

\item{top_n}{top n features selected}
}
\value{
Feature Selection Results
}
\description{
Select feature with importance
}
\author{
Wei Jiang <jiangwei@xbiome.cn>
}
