% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xbiome_jw_function.R
\name{jaccard}
\alias{jaccard}
\title{Jaccard Distance of Samples}
\usage{
jaccard(x)
}
\arguments{
\item{x}{OTU table, with rows being samples and columns being features.}
}
\value{
distance matrix
}
\description{
Give the OTU table, return the jaccard distance matrix
}
\author{
Wei Jiang <jiangwei@xbiome.cn>
}
