% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xbiome_jw_function.R
\name{lr_model}
\alias{lr_model}
\title{Pernalized Logistic Regression}
\usage{
lr_model(train, test, n = 100, method = "cv", fold = 3)
}
\arguments{
\item{train}{train SET}

\item{test}{test SET}

\item{n}{seed}

\item{method}{method for tuning hyperprameters}

\item{fold}{the number of folds used for cross validation}
}
\value{
a list containing model, confusion matrix, and test AUC
}
\description{
Apply Logistic Regression for trainning in train set and test in test set
}
\author{
Wei Jiang <jiangwei@xbiome.cn>
}
