% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xbiome_jw_function.R
\name{permanova_test}
\alias{permanova_test}
\title{Permanova Test}
\usage{
permanova_test(d, meta, feature, metric)
}
\arguments{
\item{d}{OTU TABLE with rows being sample and columns being features}

\item{meta}{metadata}

\item{feature}{a column name in meta}

\item{metric}{a distance metric used to compute the distance between samples}
}
\value{
a statistac res, p value is get by res$aov.tab$`Pr(>F)`[1]
}
\description{
Permanova test used to test the dispersion difference
}
\author{
Wei Jiang <jiangwei@xbiome.cn>
}
