% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xbiome_jw_function.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{Plot Heatmap of OTU table}
\usage{
plot_heatmap(data, feature, label)
}
\arguments{
\item{data}{an otu table}

\item{feature}{feature used in otu table}

\item{lable}{a column in data used to plot color bar}
}
\description{
Plot Heatmap of OTU table
}
\author{
Wei Jiang <jiangwei@xbiome.cn>
}
