% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xbiome_jw_function.R
\name{plot_nmds}
\alias{plot_nmds}
\title{Non-metric Multi-Dimensional Analysis(NMDS)}
\usage{
plot_nmds(data, meta, column)
}
\arguments{
\item{data}{OTU table with rows being sample names, and columns being features}

\item{meta}{meta table, with rows being sample names}

\item{column}{a scalar, quoted, denoted which environment factor should be plot with color}
}
\description{
plot Non-metric Multi-Dimensional Analysis results
}
\examples{
fun(1, 1)
fun(10, 1)
}
\author{
Wei Jiang <jiangwei@xbiome.cn>
}
