% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xbiome_jw_function.R
\name{plot_pca}
\alias{plot_pca}
\title{Principal Component Analysis}
\usage{
plot_pca(data, meta, column)
}
\arguments{
\item{data}{OTU Table with rows being samples and columns being feature;}

\item{meta}{meta table}

\item{column}{a column in meta table used as reference}
}
\description{
Give otu table and meta table to generate PCA plot
}
\author{
Wei Jiang <jiangwei@xbiome.cn>
}
