% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xbiome_jw_function.R
\name{plot_pcoa}
\alias{plot_pcoa}
\title{Principal Cordinate Analysis}
\usage{
plot_pcoa(data, meta, column, metric, trefile = "")
}
\arguments{
\item{data}{OTU TABLE with rows being sample and columns being features}

\item{meta}{metadata}

\item{column}{a column in metadata used to add color}

\item{metric}{the distance metric used to calculate the distance between samples,
including "bray", "jaccard", "unifrac", "wunifrac"}

\item{trefile}{tree file, .nwk or .tre file showing the phylogenetic tree of
features, this has to be specified if the metric is "unifrac" or "wunifrac".}
}
\value{
A list contain pcoa results(pcoa), data frame merged by pcoa and meta(data_ord), pcoa plot
}
\description{
This function is used to plot PCoA graph of otu table
}
\author{
Wei Jiang <jiangwei@xbiome.cn>
}
