% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xbiome_jw_function.R
\name{plot_pcoa_plus_seq}
\alias{plot_pcoa_plus_seq}
\title{PCoA plot with sequence Distribution}
\usage{
plot_pcoa_plus_seq(pcoa, otu_df, Response, is_sequence = TRUE)
}
\arguments{
\item{pcoa}{a pcoa object generate from plot_pcoa function}

\item{otu_df, }{the otu table used to calculate the correlation between PC1 component
and sequence feature across all samples. Relative abundant otu table is advised.
Must be a data frame.}

\item{Response}{environment variable used as color reference in PCoA graph. No quote!!!!}

\item{is_sequence, }{whether the colnames of otu table is sequence or not, if TRUE, taxonomy
is conducted using assignTaxonomy function from dada2 package}
}
\value{
Taxonoy Results is displayed
}
\description{
This function is used to combine PCoA graph and the distribution plot of
highly related sequence in PC1 axis
}
\author{
Wei Jiang <jiangwei@xbiome.cn>
}
