% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xbiome_jw_function.R
\name{plot_pcoa_with_shape}
\alias{plot_pcoa_with_shape}
\title{Principal Cordinate Analysis with 2 Confounding Factors}
\usage{
plot_pcoa_with_shape(data, meta, column, col_control, metric,
  trefile = "")
}
\arguments{
\item{data}{OTU TABLE with rows being sample and columns being features}

\item{meta}{metadata}

\item{column}{column in meta table used to plot color}

\item{col_control}{column in meta table used to plot shape}

\item{metric}{distance measure to compute the distance between samples}

\item{trefile}{tree file, .nwk or .tre file showing the phylogenetic tree of
features, this has to be specified if the metric is "unifrac" or "wunifrac"}
}
\description{
This function is used to plot PCoA with 2 variables, display in graph as color and shape
}
\author{
Wei Jiang <jiangwei@xbiome.cn>
}
