% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xbiome_jw_function.R
\name{plot_roc}
\alias{plot_roc}
\title{ROC curve}
\usage{
plot_roc(model, train_selected)
}
\arguments{
\item{model}{a model after trained}

\item{rain_selected}{training dataset}
}
\description{
plot ROC curve in train set
}
\author{
Wei Jiang <jiangwei@xbiome.cn>
}
