% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xbiome_jw_function.R
\name{plot_test_roc}
\alias{plot_test_roc}
\title{Test ROC}
\usage{
plot_test_roc(model, test_data)
}
\arguments{
\item{model}{a model after trained}

\item{test_selected}{test dataset}
}
\value{
the AUC value in test set
}
\description{
plot ROC curve in test set
}
\author{
Wei Jiang <jiangwei@xbiome.cn>
}
