% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xbiome_jw_function.R
\name{pre_process}
\alias{pre_process}
\title{Process as Merged Paired End}
\usage{
pre_process(seqRes_double_exp, trim_len)
}
\arguments{
\item{seqRes_double_exp}{output from dada2 results}

\item{trim_len}{reads whose length is less than this value will be dumped}
}
\value{
A list contain Abd and Rel_Abd
}
\description{
This function is used to to process results from dada2 results,
which will trim the sequence with length trim_len from the end of
reverse reads as features. Mostly the same with process_se, but this
function can trim more than single reads.
}
\author{
Wei Jiang <jiangwei@xbiome.cn>
}
