% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xbiome_jw_function.R
\name{process_se}
\alias{process_se}
\title{Process Single Reads}
\usage{
process_se(seqRes_double_exp, trim_len)
}
\arguments{
\item{seqRes_double_exp}{output from dada2 results, list object containing}

\item{trim_len}{the length of features that will be trimmed from, the end of reverse reads. Technically, this value should be no more
than 217(or the length of single reads)}
}
\value{
An otu table with rows being samples and columns being features.
}
\description{
This function is used to process output results from dada2,
which will trim the sequence with length trim_len from the end of
reverse reads as features.
}
\author{
Wei Jiang <jiangwei@xbiome.cn>
}
