% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xbiome_jw_function.R
\name{rarefy_otu}
\alias{rarefy_otu}
\title{Rarefaction of Otu Table}
\usage{
rarefy_otu(otu)
}
\arguments{
\item{otu}{Absolute count}
}
\value{
rarefied otu
}
\description{
Rarefy the original Otu table with minimal sample reads
}
\author{
Wei Jiang <jiangwei@xbiome.cn>
}
