% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xbiome_jw_function.R
\name{read_to_genus}
\alias{read_to_genus}
\title{Group Sequence to Genus Level}
\usage{
read_to_genus(taxa, otu)
}
\arguments{
\item{taxa, }{a data frame/matrix with row names being sequece consistent
with the column name of otu table.}

\item{otu, }{otu table with rows being samples and columns being sequence}
}
\value{
An otu table with rows being samples and columns being Genus names
}
\description{
This function is used to group feature sequence into Genus level,
Sequence with same Genus taxonomy will be add together.
}
\author{
Wei Jiang <jiangwei@xbiome.cn>
}
