% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xbiome_jw_function.R
\name{remove_batch}
\alias{remove_batch}
\title{Remove Batch Effect}
\usage{
remove_batch(otu_count, batch)
}
\arguments{
\item{otu_count}{otu table with absolute count}

\item{batch}{factor or vector indicating batches}
}
\value{
otu table de-batch-effected
}
\description{
tu table with column being feature and rows being sample;
replace zeros with half the minimal frequency;
method with removeBatchEffect from limma;
}
\author{
Wei Jiang <jiangwei@xbiome.cn>
}
