% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xbiome_jw_function.R
\name{rf_model}
\alias{rf_model}
\title{Random Foret Model}
\usage{
rf_model(train, test, n)
}
\arguments{
\item{train}{train SET}

\item{test}{test SET}

\item{n}{seed}
}
\value{
a list containing model, confusion matrix, true predicting case and test AUC
}
\description{
apply Random Forest Model for trainning in train set and test in test set
}
\author{
Wei Jiang <jiangwei@xbiome.cn>
}
