% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xbiome_jw_function.R
\name{rfe_select}
\alias{rfe_select}
\title{Feature Selection by rfe}
\usage{
rfe_select(dataset, n)
}
\arguments{
\item{n}{seed}

\item{train}{train SET}
}
\value{
Feature Selection Results
}
\description{
Apply rfe in random forest for feature selection
}
\author{
Wei Jiang <jiangwei@xbiome.cn>
}
