% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xbiome_jw_function.R
\name{write_biom_from_otu_table}
\alias{write_biom_from_otu_table}
\title{Write Biom File}
\usage{
write_biom_from_otu_table(otu, rep_fp, biom_fp)
}
\arguments{
\item{otu}{otu table for write}

\item{rep_fp}{file path to write representative sequence to}
}
\description{
write otu table into a biom file
}
\author{
Wei Jiang <jiangwei@xbiome.cn>
}
