% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xbiome_jw_function.R
\name{xgb_model}
\alias{xgb_model}
\title{XGBoost Model}
\usage{
xgb_model(train, test, n)
}
\arguments{
\item{train}{train SET}

\item{test}{test SET}
}
\value{
a list containing model, confusion matrix
}
\description{
Apply XGB Model for trainning in train set and test in test set
}
\author{
Wei Jiang <jiangwei@xbiome.cn>
}
